BlockChain Code/requirement.txt
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pip==9.0.1
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
requests==2.22.0
setuptools==39.0.1
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
